% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{ccaa_iso}
\alias{ccaa_iso}
\title{ISO code of Spanish regions}
\format{
An object of class \code{data.frame} with 20 rows and 3 columns.
}
\source{
\url{https://es.wikipedia.org/wiki/ISO_3166-2:ES}
}
\usage{
ccaa_iso
}
\description{
The ISO 3166-2 codes for Spanish regions (autonomous communities)

\describe{
\item{label}{The autonomous community name, as it appears in INE datasets.}
\item{iso}{The ISO code of the autonomous community (e.g., ES-CT for Catalonia).}
\item{nombres}{The official name of the autonomous community.}
}
}
\examples{

ccaa_iso

}
\keyword{datasets}
