% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_aquaprobe_data.R
\name{compile_aquaprobe_data}
\alias{compile_aquaprobe_data}
\title{Compile aquaprobe data}
\usage{
compile_aquaprobe_data(path_in, path_out, export = TRUE, verbose = TRUE)
}
\arguments{
\item{path_in}{A full file path to the directory containing parsed .csv files.}

\item{path_out}{A full file path to the location where output files are to be written.}

\item{export}{Logical indicating whether to save a csv of the compiled data in \code{path_out}}

\item{verbose}{Logical indicating whether to print messages.}
}
\value{
data.frame
}
\description{
This function consumes .csv files produced by the \code{parse_aquaprobe_xlsx} and compiles them into a single data set.
}
\examples{
compile_aquaprobe_data(path_in = "/User/test/aquaprobe/raw/csv/input.csv",
                       path_out = "/User/test/aquaprobe")

}
