% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_tac_data.R
\name{compile_tac_data}
\alias{compile_tac_data}
\title{Compile TAC data}
\usage{
compile_tac_data(path_in, path_out, tau, key, verbose = TRUE)
}
\arguments{
\item{path_in}{A full file path to the directory containing parsed .csv files.}

\item{path_out}{A full file path to the location where output files are to be written.}

\item{tau}{The threshold Ct value for the amplification control. For each TAC, when relevant amplification controls (e.g. MS2, PhHV) are BELOW this threshold, "Undetermined" responses can be set to this threshold value.}

\item{key}{A data.frame containing the mapping of unique and concise target names and amplification controls. See \code{esdata::key_dhaka} for template.}

\item{verbose}{Logical indicating whether to print messages. Default set to TRUE.}
}
\value{
data.frame
}
\description{
This function consumes .csv files produced by the \code{parse_taq_eds} and \code{parse_taq_xls} and compiles them into a single data set. When compiling, duplicated names are corrected as defined by the data in the
\code{key} data set included in this package. Adjustments to the Ct values are made using blank samples to control for contamination and relevant RNA/DNA targets as qPCR amplification controls.
}
\examples{
\dontrun{
compile_tac_data(path_in = "/User/test/rtac/raw/csv",
                 path_out = "/User/test/rtac",
                 key = esdata::key_dhaka)
}
}
