% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_date_time.R
\name{convert_date_time}
\alias{convert_date_time}
\title{Convert Aquaprobe data and time to consistent date format}
\usage{
convert_date_time(x)
}
\arguments{
\item{x}{A vector containing character strings of the 'Date & Time' column from an Aquaprobe data set}
}
\value{
vector
}
\description{
This function takes a vector of the 'Date & Time' column in the Aquaprobe data, and transforms it into a
consistent Date-class object. Note that many of the dates in this column are corrupted and give NAs.
}
\examples{
\dontrun{
convert_date_time(df$`Date & Time`)
}
}
