% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_dms_to_dd.R
\name{convert_dms_to_dd}
\alias{convert_dms_to_dd}
\title{Convert Aquaprobe coordinates from DMS to DD}
\usage{
convert_dms_to_dd(x)
}
\arguments{
\item{x}{A vector containing character strings of Latitude or Longitude in the informal DMS format from the Dhaka aquaprobe .xlsx files.}
}
\value{
List with both DMS and DD formats
}
\description{
This function takes the character strings giving Latitude and Longitude from the Aquaprobe .xlsx files, and returns spatial coordinates in
Degree Minute Second (DMS) and Decimal Degree (DD) formats.
}
\examples{
\dontrun{
convert_dms_to_dd(df$Lat)
}
}
