% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello2.R
\docType{data}
\name{defunciones}
\alias{defunciones}
\title{weekly report of deaths in Spain by province}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 148632 rows and 4 columns.
}
\source{
\url{https://www.ine.es/jaxiT3/files/t/es/csv_bdsc/35176.csv?nocab=1}
}
\usage{
defunciones
}
\description{
Number of reported deaths for each week for each province of Spain

\describe{
\item{year}{Year of datum.}
\item{semana}{Number of week of datum.}
\item{region}{The ISO 3166-2 code for the province of each row. Spanish data have the 'ES' label here.}
\item{valor}{The value of the datum for each row.}
}
}
\examples{

defunciones

}
\keyword{datasets}
