% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello2.R
\docType{data}
\name{defunciones_sexo_edad}
\alias{defunciones_sexo_edad}
\title{weekly report of deaths in Spain by province, gender and age group}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 442260 rows and 6 columns.
}
\source{
\url{https://www.ine.es/jaxiT3/files/t/es/csv_bdsc/36166.csv?nocab=1}
}
\usage{
defunciones_sexo_edad
}
\description{
Number of reported deaths for each week for each province of Spain, split by gender and age group

\describe{
\item{year}{Year of datum.}
\item{semana}{Number of week of datum.}
\item{region}{The ISO 3166-2 code for the province of each row. Spanish data have the 'ES' label here.}
\item{sexo}{The gender for each row: `total' for the sum of both genders, 'hombres' for men and 'mujeres' for women.}
\item{edad}{The age segment for each row.}
\item{valor}{The value of the datum for each row.}
}
}
\examples{

defunciones_sexo_edad

}
\keyword{datasets}
