% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{epa_form}
\alias{epa_form}
\title{Encuesta de Población Activa (Labor Force Survey) by region, training and gender}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 38880 rows and 6 columns.
}
\source{
\url{https://www.ine.es/dynt3/inebase/es/index.htm?padre=982&capsel=983}
}
\usage{
epa_form
}
\description{
The Spanish Labor Force Survey by region, training and gender in tidy format.

\describe{
\item{periodo}{The last day of the quarter of each row, in date format.}
\item{region}{The ISO 3166-2 code for the region of each row. Spanish data have the 'ES' label here.}
\item{form}{The traning level for each row: 'total' (all sample), 'analf' (illiterate), 'prim_inic' (started primary education), 'prim' (primary education completed), 'sec_1' (first level of secondary education completed), 'sec_2_gen' (second level of secondary education completed, general), 'sec_2_prof' (second level of secondary education completed, vocational), 'ed_sup' (higher education).}
\item{sexo}{The gender for each row: `total' for the sum of both genders, 'hombres' for men and 'mujeres' for women.}
\item{dato}{The type of data for each row: 'pob' is active population, 'act' is total workforce, 'ocu' is employed population, 'par' is unemployed population, 'ina' is non-active population.}
\item{valor}{The value of the data for the row. All data are thousands of people.}
}
}
\examples{

epa_form

}
\keyword{datasets}
