% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{epa_nac}
\alias{epa_nac}
\title{Encuesta de Población Activa (Labor Force Survey) by region, nationality and gender}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 67500 rows and 6 columns.
}
\source{
\url{https://www.ine.es/dynt3/inebase/es/index.htm?padre=982&capsel=983}
}
\usage{
epa_nac
}
\description{
The Spanish Labor Force Survey by nationality (Spanish, EU, non-EU), training and gender in tidy format.

\describe{
\item{periodo}{The last day of the quarter of each row, in date format.}
\item{region}{The ISO 3166-2 code for the region of each row. Spanish data have the 'ES' label here.}
\item{nac}{Nationality for each row: 'total' (all sample),, 'ES' (Spanish), 'EX' (non-Spanish), 'UE' (countries of member of the European Union), 'no-UE' (countries outside the European Union).}
\item{sexo}{The gender for each row: `total' for the sum of both genders, 'hombres' for men and 'mujeres' for women.}
\item{dato}{The type of data for each row: 'pob' is active population, 'act' is total workforce, 'ocu' is employed population, 'par' is unemployed population, 'ina' is non-active population.}
\item{valor}{The value of the data for the row. All data are thousands of people.}
}
}
\examples{

epa_nac

}
\keyword{datasets}
