% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{ipc_clas_grupos}
\alias{ipc_clas_grupos}
\title{Classification levels for goods to compute the Spanish Consumer Price Index}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 464 rows and 4 columns.
}
\source{
\url{https://www.ine.es/dynt3/inebase/es/index.htm?padre=3470&capsel=3466}
}
\usage{
ipc_clas_grupos
}
\description{
Spanish names of the groupings of goods use to compute the Consumer Price Index in Spain.

\describe{
\item{label}{The classification level, as it appears on the original INE datasets.}
\item{nivel}{From larger to smaller level of aggregation we have'grupos', 'subgrupos', 'clases' and 'subclases'. 'rubricas' and 'grupos_especiales' are alternative classifications.}
\item{codigo}{The label used in the 'grupo' variable in the datasets of Spanish Consumer Price Index. The global Consumer Price Index has the code 'general'.}
\item{nombre}{The name of the classification level, edited from the original names in the original INE datasets.}
}
}
\examples{

ipc_clas_grupos

}
\keyword{datasets}
