% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{ipc_valores}
\alias{ipc_valores}
\title{Names and labels of data provided in the Consumer Price Index databases}
\format{
An object of class \code{data.frame} with 4 rows and 2 columns.
}
\source{
\url{https://www.ine.es/dynt3/inebase/es/index.htm?padre=3470&capsel=3466}
}
\usage{
ipc_valores
}
\description{
Names and labels of data provided in the Consumer Price Index databases.

\describe{
\item{dato}{The Spanish name of each type of data.}
\item{dato_label}{The label used in the 'dato' variable of the Consumer Price Index datasets.}
}
}
\examples{

ipc_valores

}
\keyword{datasets}
