% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{ipv}
\alias{ipv}
\title{Spanish Housing Price Index by autonomous community}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13888 rows and 5 columns.
}
\source{
\url{https://www.ine.es/dyngs/INEbase/es/operacion.htm?c=Estadistica_C&cid=1254736152838&menu=ultiDatos&idp=1254735976607}
}
\usage{
ipv
}
\description{
Quarterly Housing Price Index (Índice de Precios de la vivienda) of Spain and autonomos communities.

\describe{
\item{periodo}{The last day of the qaurter of each row, in date format.}
\item{region}{The ISO 3166-2 code for the region of each row. Spanish data have the 'ES' label here.}
\item{tipo}{The label 'general' is for the generic housing prices, 'nueva" for new housing, and 'segunda_mano' for second-hand housing.}
\item{dato}{The type of datum for each row: 'index' is the Housing Price Index,  'mensual' is the monthly variation of the index, 'anual' the inter-annual variation of the index and 'acumulada' the cumulative variation of the index for the year.}
\item{valor}{for rows with dato equal to 'indice' the Jousing Price Index,  housing price variation in percentage for the rest of values.}
}
}
\examples{

ipv

}
\keyword{datasets}
