% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{ipv_pond}
\alias{ipv_pond}
\title{Weights of groups in the Housing Price Index by region and year}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 468 rows and 4 columns.
}
\source{
\url{https://www.ine.es/dynt3/inebase/es/index.htm?padre=3470&capsel=3466}
}
\usage{
ipv_pond
}
\description{
Weights of new and seconds hand house in the Housing Price index by region and year. The sum of weights for a region and year is equal to 100.

\describe{
\item{periodo}{The year of the weight datum, in numeric format.}
\item{region}{The ISO 3166-2 code for the region of each row. Spanish data have the 'ES' label here.}
\item{tipo}{The label 'nueva" si for new housing, and 'segunda_mano' for second-hand housing.}
\item{valor}{The value of the weight, ranging from 0 to 100.}
}
}
\examples{

ipv_pond

}
\keyword{datasets}
