% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_aquaprobe_xlsx.R
\name{parse_aquaprobe_xlsx}
\alias{parse_aquaprobe_xlsx}
\title{Parse Aquaprobe .xlsx files}
\usage{
parse_aquaprobe_xlsx(path_in, path_out, verbose = TRUE)
}
\arguments{
\item{path_in}{A full file path to the directory containing the .xlsx files.}

\item{path_out}{A full file path to the location where output files are to be written.}

\item{verbose}{Logical indicating whether to print messages.}
}
\value{
data.frame
}
\description{
This function takes a .xlsx file path (or vector of file paths) and extracts its contents into individual csv files containing raw data.
}
\examples{
\dontrun{
parse_aquaprobe_xlsx(path_in = c("/User/test/dir1", "/Users/test/dir2"),
                     path_out = "/User/test")
}
}
