% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tac_eds.R
\name{parse_tac_eds}
\alias{parse_tac_eds}
\title{Parse Taqman .eds files}
\usage{
parse_tac_eds(path_in, path_out = NULL)
}
\arguments{
\item{path_in}{A full file path to the directory containing the .eds files.}

\item{path_out}{A full file path to the location where output files are to be written.}
}
\description{
This function takes a directory with .eds files produced by QuantStudio software and extracts its contents into xml and txt files containing raw data and experiment metadata.
}
\examples{
\dontrun{
parse_tac_eds(path_in = "/User/test/dir1",
              path_out = "/User/test")
}
}
