% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tac_xls.R
\name{parse_tac_xls}
\alias{parse_tac_xls}
\title{Parse Taqman .xls files}
\usage{
parse_tac_xls(path_in, path_out, verbose = TRUE)
}
\arguments{
\item{path_in}{A full file path to the directory containing the .xls files.}

\item{path_out}{A full file path to the location where output files are to be written.}

\item{verbose}{Logical indicating whether to print messages.}
}
\value{
data.frame
}
\description{
This function takes a directory with .xls files produced by QuantStudio software and extracts its contents into individual csv files containing raw data.
}
\examples{
\dontrun{
parse_tac_xls(path_in = c("/User/test/dir1", "/Users/test/dir2"),
              path_out = "/User/test")
}
}
