% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{pib_demanda_agregados}
\alias{pib_demanda_agregados}
\title{Aggregates of the Gross Domestic Product through the expenditure approach}
\format{
An object of class \code{data.frame} with 36 rows and 2 columns.
}
\source{
\url{https://www.ine.es/dynt3/inebase/es/index.htm?padre=5614&capsel=5615}
}
\usage{
pib_demanda_agregados
}
\description{
The code used to represent the GDP aggregates in 'pib_pm_demanda' and the name of each aggregate as presented in original data.

\describe{
\item{agr_label}{The label used in the 'agregado' field of 'pib_pm_demanda'.}
\item{agregado}{The actual name of the aggregate as presented in original data (in Spanish).}
}
}
\examples{

pib_demanda_agregados

}
\keyword{datasets}
