% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{pib_oferta_agregados}
\alias{pib_oferta_agregados}
\title{Aggregates of the Gross Domestic Product through the production  approach}
\format{
An object of class \code{data.frame} with 16 rows and 3 columns.
}
\source{
\url{https://www.ine.es/dynt3/inebase/es/index.htm?padre=5614&capsel=5615}
}
\usage{
pib_oferta_agregados
}
\description{
The code used to represent the GDP aggregates in 'pib_pm_oferta' and the name of each aggregate as presented in original data.

\describe{
\item{agr_label}{The label used in the 'agregado' field of 'pib_pm_oferta'.}
\item{agregado}{The actual name of the aggregate as presented in original data (in Spanish).}
\item{nivel}{Level of the aggregate.}
}
}
\examples{

pib_oferta_agregados

}
\keyword{datasets}
