% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{pib_pm_oferta}
\alias{pib_pm_oferta}
\title{Gross domestic product (GDP) of Spain, production approach}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 40552 rows and 7 columns.
}
\source{
\url{https://www.ine.es/dynt3/inebase/es/index.htm?padre=5614&capsel=5615}
}
\usage{
pib_pm_oferta
}
\description{
Quearterly Gross Domestic product of spain, calculated through the production apprach. Includes the final value of GDP and its aggregates. Includes indices and total value for each aggregate. Data from 1995 to present.

\describe{
\item{periodo}{The last day of the quarter of each row, in date format.}
\item{agregado}{The code for each aggregate of the GDP through the production approach. See 'pib_oferta_agregados' for a list of aggregates (in Spanish).}
\item{nivel}{Level of the aggregate.}
\item{ajuste}{'ajustad' means that data are adjusted from seasonality and calendar, and 'no_ajustado' means that are not adjusted.}
\item{tipo}{Rows with 'valor' contain the actual GDP value, and rows with 'indice' and indexed value.}
\item{dato}{The type of datum for each row. 'base' means the actual value, 'var_trim' the quarterly variation, 'var_anual' the annual variation, 'ap_trim' the quarterly contribution and 'ap_anual' the annual contribution.}
\item{valor}{The value of the datum for each row.}
}
}
\examples{

pib_pm_oferta

}
\keyword{datasets}
