% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello2.R
\docType{data}
\name{pob_prov_edad}
\alias{pob_prov_edad}
\title{Spanish population by region (province), age and gender}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 80454 rows and 5 columns.
}
\source{
\url{https://www.ine.es/jaxi/files/_px/es/csv_bdsc/t20/e245/p08/l0/03002.csv_bdsc?nocab=1}
}
\usage{
pob_prov_edad
}
\description{
The Spanish population by region, age and gender. Updated yearly.

\describe{
\item{periodo}{The year of the datum.}
\item{region}{The ISO 3166-2 code for the province of each row. Spanish data have the 'ES' label.}
\item{edad}{The age segment for each row.}
\item{sexo}{The gender for each row: `total' for the sum of both genders, 'hombres' for men and 'mujeres' for women.}
\item{valor}{The value of the datum of each row..}
}
}
\examples{

pob_prov_edad

}
\keyword{datasets}
