% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello2.R
\docType{data}
\name{prov_ccaa}
\alias{prov_ccaa}
\title{ISO codes of provinces and corresponding region}
\format{
An object of class \code{data.frame} with 53 rows and 2 columns.
}
\usage{
prov_ccaa
}
\description{
The ISO codes of each province, linked with the ISO code of its region. Can be useful to get regional data from provincial data.

\describe{
\item{iso_prov}{The ISO 3166-2 code for the province of each row.}
\item{iso_ccaa}{The ISO code of the region (autonomous community).}
}
}
\examples{

prov_ccaa

}
\keyword{datasets}
