% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{provincias_iso}
\alias{provincias_iso}
\title{ISO code of Spanish provinces}
\format{
An object of class \code{data.frame} with 53 rows and 3 columns.
}
\source{
\url{https://es.wikipedia.org/wiki/ISO_3166-2:ES}
}
\usage{
provincias_iso
}
\description{
The ISO 3166-2 codes for Spanish provinces

\describe{
\item{label}{The autonomous community name, as it appears in INE datasets.}
\item{iso}{The ISO code of the autonomous community (e.g., ES-B for Barcelona).}
\item{nombres}{The official name of the province.}
}
}
\examples{

provincias_iso

}
\keyword{datasets}
