% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standard_curve_dhaka.R
\docType{data}
\name{standard_curves_dhaka}
\alias{standard_curves_dhaka}
\title{Standard curve data for Dhaka analyses}
\format{
\subsection{\code{standard_curves_dhaka}}{

A data frame with 4 columns:
\describe{
\item{target_name_unique}{The original target name found on each of the taqman array cards. May contain typos and duplicates.}
\item{target_name_concise}{The associated concise target name with typos and duplicates removed.}
\item{ct_value_1}{The first measured Ct value attained for the particular viral load dilution.}
\item{ct_value_2}{The second measured Ct value attained for the particular viral load dilution.}
\item{quantity}{The number of viral copies for the particular dilution.}
}
}
}
\usage{
standard_curves_dhaka
}
\description{
These data provide the standard curve relationship between viral load and the qPCR Ct value. The data pertain only to analyses
run for samples from Dhaka, Bangladesh. However, the format of these data can be replicated when modeling viral load for other data sources.
}
\keyword{datasets}
