% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\docType{data}
\name{ipc_hist_grupos}
\alias{ipc_hist_grupos}
\title{Historical series of inflation rates by group (since 1993)}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 12769 rows and 4 columns.
}
\source{
\url{https://www.ine.es/dynt3/inebase/es/index.htm?padre=3470&capsel=3466}
}
\usage{
ipc_hist_grupos
}
\description{
@description Monthly inflation data of Spain by group of goods since 1993.
}
\details{

}
\examples{

ipc_hist_grupos

}
\keyword{datasets}
